Added cf calculation

+++ DGP_code/run.sh
@@ -1,15 +1,16 @@
 #!/bin/bash
 
 PROCS=$1
-rm -rf RUN
+DELETE=$2
+
+if [ $DELETE -gt 0 ]; then
+    echo "New case. Deleting old directories"
+    rm -rf RUN
+    mkdir RUN;
+fi
 echo "Running Job on $PROCS processors"
 echo "See RUN folder"
-mkdir RUN
 cp control.txt RUN/
 cd RUN
 mpirun -np $PROCS ../bin/DGPLS
 
-cd ..
-rm -r RUN0
-mkdir RUN0
-cp -r RUN/* RUN0/
